feat: Перенос функционала в Server, удаление валидации из Server
